Add test for removing generated enum methods and fields
Summary:
We want to remove the generated static methods `Enum.valueOf()` and `Enum.values()` and the generated private field `$VALUES`. There are two reasons why we wouldn't be able to remove them.
1. The method is called directly, e.g., `MyEnum.valueOf("ONE")`.
2. The method is called reflectively, e.g., `Class.getDeclaredMethod("valueOf")`. See [Enum.java](https://android.googlesource.com/platform/libcore/+/refs/heads/master/ojluni/src/main/java/java/lang/Enum.java#267).

We can detect (2) by checking if the enum is ever upcasted to either an `Object` type or another super type and if it escapes a method. We also need to check if the enum ever escapes as the type `java.lang.Class` by checking for `Enum.getDeclaredClass()`.

Reviewed By: int3

Differential Revision: D16009620

fbshipit-source-id: 8d043fdce2c235bdf0309b6650bbeb3ebbc17705