make EXPECT_CODE_EQ macro show diff
Summary: After discussing with Jez, we decided to focus first on a differential indicator rather than pretty printing, as it did not seem like a helpful pretty printing functionality was practically possible. This diff adds the functionality to indicate where the difference between two inputs start when the EXPECT_CODE_EQ macro is used.

Reviewed By: int3

Differential Revision: D16086514

fbshipit-source-id: 67f3033ea5e340885aec7b26b49b17f2dd86cda0