Reenable Constant Propagation for Switch
Summary:
This patch reenables constant propagation for switch which caused a test failure.
The feature tried to remove aggressively any redundant labels in default block for a switch.
In the failing case, there is only a single/default block left without any switch targeted label while leaving the statement, which resulted in an inconsistent control flow graph. The fix is to simply remove the switch statement as well in this case.
A test case is also added to cover this case.

Reviewed By: justinjhendrick

Differential Revision: D16097243

fbshipit-source-id: 2c86c602d172ce3cdc9b02742bb595c261d2b762