fix infinite loop case
Summary:
Fixes https://github.com/facebook/redex/issues/365

This also changes the behavior of the CFG inliner around merging blocks with single a goto edge between them. Previously, the CFG inliner would merge the blocks in this scenario but now it won't. Doing this optimization inside the inliner made it overly complicated and it's the same thing as RemoveGotosPass. So, we can use that pass after the inliner if we need this behavior in the future

Reviewed By: NTillmann

Differential Revision: D15441763

fbshipit-source-id: 918b57919309ee2a029166c5436f07488226e0b8