Fix some usages of Editable CFG
Summary:
When using the editable CFG:
(1) You must deconstruct at the end of your pass. Currently, passes assume the incoming code will be in IRCode form
(2) You can't use the IRCode while the editable CFG is built

This diff fixes some occurences of these bad patterns and enforces (1) going forward. I'd like to enforce (2) but it's not ready yet, see D13968654 for more details on that

Reviewed By: Feng23

Differential Revision: D16213368

fbshipit-source-id: ac20cbbd8717c23f172d143cddd2d14fa8cb286b