Refactor to write to ostream for the sake of testing
Summary:
We want to test to verify the output of IODIMetadata is correct & valid. In order to test this without actually writing to disk we need to be able to pass in our own stream to write to.

Additionally I move the computation of IODI metadata's header to the start of write() since stringstream doesn't support seeking past eof (i.e. in test scenario the code would fail)

Reviewed By: int3

Differential Revision: D15850393

fbshipit-source-id: 11b1564663373bc703e13a0d5272174dca9a1d04