Introduce API to load dexes from memory
Summary: In order to make an integration test that inspects an outputted dex file we need to be able to reload the outputted dex file into our abstractions. In order to avoid creating a temporary file we need to introduce an API that avoids loading a file from disk and accepts a buffer.

Reviewed By: int3

Differential Revision: D15844563

fbshipit-source-id: 293e4eee5a0f7736926647a9e606ac7d4dd9c84a