Abstract out compute_method_to_id_map and invoke within prepare
Summary: In order to test the output data from DexOutput without writing to a file and then re-reading it we need to compute all the data that will be written before `DexOutput::write()`. With this change we can create tests that avoid writing a dex file to disk, but still allows us testing method_to_id_map's contents.

Reviewed By: int3

Differential Revision: D15850395

fbshipit-source-id: e30cad16f759ef3ba7ba9517bee08a347a178c1e