Add source data to DexDebugItem
Summary:
In order to create IODI tests we need to be able to tell if two pieces of debug info came from the same file & were at the same offset. In order to check for this, we introduce some new parameters on DexDebugItem.

Long term I think we should rethink these data structures to have a clear set of data structures that simply mirror dex data and then a set of data structures that represent IR versions (even if the IR versions are simply duplicates of the dex versions). This will enable us to parse a dex file and reuse e.g. DexDebugItem when multiple code items point to the same one. For now, this is fine.

Reviewed By: wsanville

Differential Revision: D15844562

fbshipit-source-id: 6f340856f755baea4d46468006994d0eb81e907e