Tune method/field ordering within class for (compressed) size
Summary:
Already in the past, the ObfuscatePass tried to reduce size by splitting up static fields into two sections: those with no static value, and those with static value. (This influences the number and size of required encoded arrays to hold static values in the dex size.)

This change does the following...
- The existing check on static values didn't actually do what was intended, as static values are already present for static fields, they just happen to be zero or non-zero. This is getting fixed.
- Further re-ordering is done for fields and methods to place those which are "similar" close to each other, e.g. sharing the same static value/proto/type/access. This again reduces the number of needed encoded static values for static fields, but more importantly, it increases the compressibility of associated metadata structures. The re-ordering happens before renaming, after which all tables are again in sorted-by-renamed-names order, while the underlying items got reshuffled.

The main intended (and actual) impact here is to reduce compressed size, and just a bit uncompressed size.

Reviewed By: justinjhendrick

Differential Revision: D16188760

fbshipit-source-id: ec002782fa1e2b7e927a78d3138ea7e0b0ccc2d0