Analyze enums for removing Enum.valueOf and Enum.values
Summary:
We want to remove the generated static methods `SubEnum.valueOf()` and `SubEnum.values()`. We cannot use `RemoveUnreachablePass` to remove them because `Class.isEnum()` and `Class.getEnumConstants()` which use these methods are commonly used in (de)serialization libraries so that we need Proguard rule to keep these methods. There are two reasons why we wouldn't be able to remove them.
1. The method is called directly, e.g., `SubEnum.valueOf("ONE")`.
2. The method is called reflectively, e.g., `Class.getDeclaredMethod("...")`    and `Class.getEnumConstants()`. See [Enum.java](https://android.googlesource.com/platform/libcore/+/refs/heads/master/ojluni/src/main/java/java/lang/Enum.java#267).

We can detect (2) by checking if the enum is ever upcasted to either an `Object` type or an `Enum` type and if it escapes a method so that it becomes hard to track. This could happen when an upcasted enum is returned from a method, set to a class variable or array, or passed as an argument to a method. We also need to check if the enum is ever used as the type `java.lang.Class` by checking for `Class.getDeclaredClass()` and `const-class`.

We also have a list of whitelisted methods that we know won't call `Enum.valueOf()` or `Enum.values()` either directly or reflectively.

Reviewed By: Feng23

Differential Revision: D16134474

fbshipit-source-id: ae7030f92c4a8661aa5db70be6c8b91142af93a0