safely eliminate many const-literals
Summary:
Background:
Code contains many seemingly redundant const-instructions.
However, the Android verifier checks how const values are used, and it will reject code that uses the same register inconsistently along any execution path, where inconsistently means that the register is used with conflicting type categories, e.g. once as an `int`, and then again as a `float`.
At a high level, the Android verifier considers three type categories for <=32-bit values: `int`, `float`, `object`. All smaller integer types are "implicitly" widened to `int`. And for 64-bit values, there is `long`, `double`.
Some instruction impose exact type demands, e.g. `ADD_INT` demands an `int`, not a `float`, and not an `object`. But `IF_EQZ` can be given an `int` or an `object`.

In the actual Android verifier implementation, the exact tracked types are much more involved, considering all relevant smaller integer types and their relationships. See here: https://android.googlesource.com/platform/dalvik/+/android-cts-4.4_r4/vm/analysis/CodeVerify.cpp#186 However, all that complexity in the Verifier exists to ensure that a value is never used in a context where certain bits should be 0 or 1. We can ignore these details for constants, as a constant is a constant is a constant.

Our approach:
If we have two const instructions loading the same bit patterns, we can drop one if they don't have mismatching type demands along all execution paths.
To avoid doing a potentially expensive path-sensitive analysis for each pair of const instructions, this diff applies one major simplification to the problem: For each const instruction, we look at all type demands across all execution paths, and compute the intersection of all these demands. In this way, if two const instructions have the same combined type demands, then it is safe to eliminate one of them. This is quite conservative, but safe.

Note that the combined type demands might be contradictory (represented as `TypeDemand::Error`). Consider the following instructions:

  const r0, 1234
  if-eqz r1, L1
  add-int r0, r0, r0
  goto L2
  :L1
  add-float r0, r0, r0
  :L2

My understanding is that this is verifiable, as there is no inconsistent feasible execution path. However, the path-insensitive combined type demand for `r0` is inconsistent, and thus `r0` will not be considered for elimination. This is the most extreme aspect of how this approach to the problem is safe.

A few details on how type demands are computed:
- For `return` and `return-wide` instructions, we look at the method result type.
- For the the value passed in to an `aput` or `aput-wide` we try to determine the incoming array type using `TypeInference`. (I am not sure if `TypeInference` can always decide the array type, due to some simplications that might exist in our type inference implementation. The Android verifier certain is able to always figure out the exact array type if the code verifies. Anyway, if in doubt, we can always safely bail out with `TypeDemand::Error`)

Reviewed By: int3

Differential Revision: D16271362

fbshipit-source-id: c7e02e83f96aabae796b7cdf43a59655d4640c7c