RMUF: Make bisection a bit more convenient
Summary:
We've had to bisect the list of removed fields several times now. Each time
I've done it with some ad-hoc code that didn't get committed, but I think it's
worthwhile to make it an official API if we're going to do it repeatedly.

Workflow:

1. Enable the `export_removed` flag in RMUF's pass config.
2. Look for `redex-removed-fields.txt` in the APK output directory. Copy it to some tempdir.
3. Run `bisect-file.sh redex-removed-fields.txt`, which will produce two files, `redex-removed-fields{0,1}.txt`.
4. Pass one of those files to RMUF via the `whitelist_file` config option. Test the resulting APK to see if the bug repros; if not try using the other file.
5. Repeatedly bisect the file that produced an APK reproing the bug until you have found the culprit.

I'll write this up more formally in a wiki after this lands...

Reviewed By: Feng23

Differential Revision: D16355198

fbshipit-source-id: 35bdb74e2cdbf4fd1912472d9f293bf523c1c50c