RMUF: Factor out removal policy logic
Summary:
I want to add a whitelist feature w/o more code duplication; this diff is a setup for that

I wanted to avoid having a long list of captured lambda parameters for the parallel walker job. Instead I've factored things so that the captured variables are now class members, and the parallel code lives in a `const` method, so we have the same safety guarantees for parallelism.

Reviewed By: Feng23

Differential Revision: D16294344

fbshipit-source-id: 2f4fefbfa8bfa74d6ac7c6a02b388ab27ee00740