load proguard maps to obtain original names in reflection analysis
Summary: We construct the representations of proguard maps with full method/field descriptors. This works well in a lot of cases but it's not convenient to use this existing structure when return type or field type is unknown. Since name + proto should uniquely identify a method, we should be able to drop these types in the mapping.

Reviewed By: dkgi

Differential Revision: D16274457

fbshipit-source-id: f6a68c4918f0d1c640e0b2d338c348c5d957c08b