add constant propagation analysis for instance-of check on null object
Summary:
"At run time, the result of the instanceof operator is true if the value of the RelationalExpression is not null and the reference could be cast to the ReferenceType without raising a ClassCastException. Otherwise the result is false."
https://docs.oracle.com/javase/specs/jls/se11/html/jls-15.html#jls-15.20.2
So for an instance-of, if the object checking is null, the result will be false. Add this case to constant propagation analysis.

Reviewed By: int3, yuxuanchen1997

Differential Revision: D15888520

fbshipit-source-id: 9365b64db04e2f3408ddfee765ad2dc032e179c0