Don't optimize methods with sketchy monitor handling
Summary:
The Android verifier has [a check](1) to make sure that when a monitor is being held, any potentially throwing opcode is wrapped in a try region that goes to a catch-all block. The catch-all is ostensibly responsible for releasing the monitor.

Surprisingly, the verifier only performs the catch-all check for throwing opcodes contained in *some* try region. This means that a method with non-wrapped throwing opcodes in a synchronized block will verify. However, if such a method is inlined, and its callsite was wrapped in a try region that does not have a catch-all, then we have a VerifyError! More generally, any code-relocating optimizations could trigger this issue. Therefore, to be safe, we mark all methods with these non-wrapped throwing opcodes in monitor regions as don't-inline as well as no-optimize.

[1]: http://androidxref.com/9.0.0_r3/xref/art/runtime/verifier/method_verifier.cc#3553

Reviewed By: wsanville

Differential Revision: D16512267

fbshipit-source-id: 2628cea606491a017a5af973e94ea0594f963732