Don't convert switch to goto if switching on a nonexact constant
Summary:
{D16019671} had a subtle bug: if the switched-upon value was not Top, and if it
only matched one case key, then constprop would convert the switch into a goto
that branched to that one case. However, if the value was a signed domain (i.e.
interval), then it could match a single case while not guaranteeing that the
case would always be the one selected at runtime.

Differential Revision: D16532778

fbshipit-source-id: 9c1cfc12eca17c2f6813fe6c235157acfcccbbb0