Implement deserialization from S-Expr to DexDebugInstructions
Summary: Add a new s-expr node `.dbg` which will be interpreted as DexDebugInstruction during parsing s-expr. Any special opcode (0x0a-0xff) must be encoded with `EMIT` literal in the parsed s-expr, and the value of special opcode following `EMIT` is encoded in decimal format. Also fix `DBG_LAST_SPECIAL` to `0xff`. This diff is the first half towrads the completion of T45979630, the part which dumps DexDebugInstruction as s-expr will be added in one following diff.

Reviewed By: danzimm

Differential Revision: D16434380

fbshipit-source-id: 112ce9e43ac5fd7f62f612d6fa932ba584307a93