Don't optimize enums that are instance fields of serializable
Summary: `SubEnum.valueOf()` or `SubEnum.values()` are not safe to remove when `SubEnum` is an instance field of a class that is serializable. If a class has `SubEnum` instance field and extends an external class, we cannot know for sure if it is serializable. We whitelist commonly used external classes that we know cannot implement `Serializable`. Using the trace I added we can find a list of external classes that cause enums to be rejected and manually check if it implements `Serializable`.

Reviewed By: emma0303

Differential Revision: D16604314

fbshipit-source-id: 02e5e3b972a2c08a6311a5bdd24e202571ac0e4e