Remove `valueOf` and `values` methods from enums if safe
Summary:
D16134474 does the analysis to make sure these methods are safe to remove, but it does not fully remove them so that we can catch any errors. This diff will fully remove these methods so that we should get a size win.

Do not ship until we are sure D16134474 has no problems!

Reviewed By: int3

Differential Revision: D16293157

fbshipit-source-id: 8fa11c1f261be7f754ad68632859dc6296c5819c