Symbolicating field types from dexdump
Summary:
Bugfix to de-obfuscate classes in type descriptors as well.

The class regex currently assumes that classes (e.g. "LX0D/7;") are preceded by whitespace only.

This doesn't account for cases such as "ILX0D7;" that can occur in types. This diff changes the regex to capture these so they de-obfuscate as well.

Reviewed By: yuxuanchen1997

Differential Revision: D16720617

fbshipit-source-id: 81a064d37daf92c04386ef7b061f48448d374906