Reject enums that could produce colliding <init> signatures
Summary:
The transform enum to int optimization replaces the `SubEnum` type with `Integer` type. This causes problems if we have two methods like
```
Foo(SubEnum a) {...}
Foo(Integer a) {...}
```

We need to go through each constructor and check for collisions that would happen if candidate enums were transformed, and reject the enums that produce that collision.

Reviewed By: Feng23

Differential Revision: D16745299

fbshipit-source-id: fb6e2ee96d97403a792fc5035471f90f745798fa