A better cost model
Summary:
Before, the inliner was only looking at, roughly, "number of instructions" in the callee to figure out whether to inline it.
Now, the inliner estimates how many code units (2 bytes) instructions will use, and various other overheads. It also considers number of callsites, and the cost of keeping a method and its metadata around.
This results in a higher effectiveness of inlining across a variety of apps.

Also: Changing default of `inline_small_non_deletables` to `true` to simplify configurations.

Reviewed By: int3

Differential Revision: D16723318

fbshipit-source-id: b96ac05cd2a2a1235a0dd9a294feb996fbcded4e