Increase precision of constant-uses
Summary:
This is a bug fix (the bug was exposed by the more aggressive inliner changes higher in the diff stack):

`if-eq` and `if-ne` require that *both* of their incoming registers *agree* on either being an object reference, or an integer.
We use type inference to see what's known about the incoming registers, and use that information to refine the demands for those `if` instructions.

Reviewed By: int3

Differential Revision: D16723311

fbshipit-source-id: d6f6c39e7f42be21addd70efd62cfe47ac557be3