Allow primitive instance fields for replace_enum_with_int
Summary:
Primitive or string instance fields in enums that are only set in `<init>` can be replaced with static methods that map the enum's `ordinal` to the field value. This is done in the same way as the `ordinal` and `name` instance fields. Then these enums are candiates for the `replace_enum_with_int` optimization pass.

There are a few requirements for this optimization.
1. Each instance field must be either a primitive or a string.
2. Each instance field must be set in `<init>`.
3. Each instance field cannot be modified outside of `<init>`.

Reviewed By: Feng23

Differential Revision: D16589343

fbshipit-source-id: 8218eba5010541072343d3ab23ed1551855cbdd4