Log error instead of assertion for method refs inannotations
Summary:
I added the assertion error to make the pass fail for pure method refs if they
seem not be updated properly. This diff removes the assertion as it was, and
adds some error log.

There is a [GitHub issue](https://github.com/facebook/redex/issues/412) related to the error.

It should be a bug from D8/R8

Reviewed By: emma0303

Differential Revision: D16800282

fbshipit-source-id: 6ae94ac5cc0c43ee931d68401d3b178545140a2d