introduce test fixture base for redex integration tests
Summary:
Redex integ tests used to be a whole lot of copypasta.

copy_pasta

This diff adds a test fixture base for such tests. We should continue to make testing easy and reduce the size of the boilerplate.

There are a few test classes this diff did not switch to the new test fixture base.

- IODI and Peephole already have a distinct structure to prevent duplicated code
- ReachableClassesTest has a "hardcoded path for OSS automake test harness"

In addition, obfuscation e2e tests seem to have a test base but not used as a fixture for its e2e tests. It may need some further work too.

Reviewed By: int3

Differential Revision: D16865320

fbshipit-source-id: d7744b609cac306b6387557ca913041f17bb6524