Support Library / AndroidX to framework api support
Summary:
NOTE: This will be used in subsequent diffs. I wanted to get this out first.

Context: In order to replace support library code with platform apis, we need a way to detect: which and minimum api / maximum api available. NOTE that I don't know of an automated way to gather this data :( and the process I did is fairly boring.

Note: that some methods were later added (hence the extra methods defined in the file).

I basically tried to map androidX classes to framework ones. TODO: Add others. The list is short.
* grep in the release builds + https://developer.android.com/reference + Android Platform tab
* Also -> P80072619 (android X classes) + P80072825 (android support classes)

Notes:
* Not complete. Will add more as we go.
* This doesn't include anything about superclasses, but that needs to be taken into account / together with Interfaces. As in, if we have a subclass, but not the rest. One example is Transition - note that I included everything that derives from it (but we can exclude those cases for now)
** On the other hand, I think this should be done at compile time - so should be ok
* Some of the androidX classes I couldn't find in the documentation :| So ... nothing for them (example androidx/core/util/DebugUtils or androidx/versionedparcelable/VersionedParcelParcel - and much others)
* And then there are some that I can't find documentation about and it existed since API 1 (like TImeUtils)

Reviewed By: int3

Differential Revision: D16834510

fbshipit-source-id: 60875831abef0b3bd8d19a856ae2944ec972f4d1