devirt: Use MOG instead of VirtualScope
Summary:
MethodOverrideGraph.h and VirtualScope.h have overlapping functionality, so I'm working on replacing the latter with the former. The MOG was originally introduced in {D9389347} and showed a large speed win on large graphs.

This diff replaces VirtualScope's `devirtualize()` with an equivalent MOG-based implementation. It's not significantly faster (`devirtualize()` wasn't that slow to begin with) but does allow us to delete a bunch of code.

Notes:
* The new function is called `get_non_true_virtuals()` which I think is a better name; `devirtualize()` seems to imply that some state is being mutated
* `get_non_true_virtuals()` simply returns all methods for which `is_true_virtual()` is false. Previously, `is_true_virtual` did not check if the given method was abstract, but `devirtualize()` does.

While doing these changes I noticed that the MethodDevirtualizerPass had a bunch of unused code, so I've deleted those too.

Reviewed By: ssj933

Differential Revision: D16803033

fbshipit-source-id: 4fbead9ebc345f93856f3b850aa8524ff2d38736