keeprules: Support negation of interfaces/enums
Summary: Previously, we didn't have proper support for keep rules that had `!interface` or `!enum` (i.e. only matched non-interfaces or non-enums). The intention was to print an error message upon encountering that sequence of tokens, but there was a bug in the code, and in practice we would just silently create a keep rule that *did* match on an interface/enum -- the opposite of what was requested.

Reviewed By: ssj933

Differential Revision: D17003673

fbshipit-source-id: 1881c9421f8038484e39491143797c2aee0c38cc