Optimize enums with true-virtual methods
Summary:
Previously we only handled non-true-virtual methods and rejected enums that had true-virtual methods. It turns out that `Enum.toString()` is the most common method that is overridden. The only other methods that can be overridden are interface methods from `Serializable`.

If `toString()` is overridden, we will replace its call sites with static methods that contain their code. If it is not overridden, we replace its call sites with `Enum.name()`, the default implementation of `Enum.toString()`.

Reviewed By: Feng23

Differential Revision: D16777953

fbshipit-source-id: 293813609d519423555379ca161ad2e74ecd55cb