Refactoring of barrier computation in CSE
Summary:
The CommonSubexpressionElimination pass computes "barrier" information for methods. The reason is that an invocation of a method may represent some kind of memory barrier. If a method is external or native, we must assume the worst, but for known methods we may be able to compute some limited set of written fields or array element types that are actually mutated, and then CSE can retain knowledge about all other memory locations while doing abstract interpretation along the instructions.

In the past, the barrier information for methods was effectively computed by first analyzing all methods with no invocations to other methods, then all methods with invocations to methods for which barriers were already computed, and so on, until a fixpoint was reached, or some iteration limit was hit.

The code implementing this was a bit cumbersome, and the whole approach had the conceptual drawback that it would just give up on any (indirectly) recursive methods.

This diff changes that.

Now, the fixpoint is computed in a different way, always augmenting the known barriers for a method, adding in any newly discovered barriers for invoked methods. (Which might lead to finding a general memory barrier, in which case we can stop augmenting additional information for that method.) We keeping track of invocation dependency information to only run the augmentation step for methods impacted by new discoveries. The result is that we can now also compute precise barrier information for recursive methods, instead of just giving up.

Reviewed By: kathryngray

Differential Revision: D16832396

fbshipit-source-id: 013450a005dda0d2d3944001c1b18ea7aec454d9