Remove the usage of predefined mixedmode classes
Summary:
As title. We never actually used this feature and we will not, so removing it for clarity.

Ps. This will also make it easier for Brian's changes from converting statically determined mixedmode dex(es) to dynamically doing this for different experiments.

Ps. Will port to staging and stable after review.

Reviewed By: newobj

Differential Revision: D17403455

fbshipit-source-id: edb4ed2ebc580ee9a3c3947555ffce58b2f2b949