Fix normal primary dex ordering
Summary:
As title. Since we started adding a lot more in the instagram betamap + the fact that we removed the end marker (DexEndMarker0) some time ago -> resulted in classes that were in the primary being ordered later on, which might have gotten then to the second dex.

We currently have an end marker (ColdStartClassPreloadingMarkerEnd), but for now, removing this condition and just blindly copying the primary dex as it is.

Reviewed By: int3

Differential Revision: D17440890

fbshipit-source-id: 58b7306213a9cceea97b24673ef5b1455024027e