let MethodInline be able to inline some simple true virtual methods
Summary:
Method Inline didn't handle true virtual before, so we are missing some opportunity like only one implementation is actually used across code base but we are keeping all the implementation, or some simple return 0 that are used in branch are not being able to be known by program because the method is true virtual.
This diff will inline methods' content into callsites that are only pointing to one definite implementation that don't have any override. **But** we will blacklist methods whose return type is not primitive or in its method body it accessed any type to avoid verification error.

Reviewed By: int3

Differential Revision: D16296014

fbshipit-source-id: 960e0048cd12d5b4a7532f6aa50e39cc810035fc