Reflect default values for bool type
Summary:
Reflect the default values we set in C++ (only `bool` for this diff) so I can get some warnings if the there are statements setting the configs to default values and it becomes easier to delete unused refig settings.
Also add a column to the docs for default values.

Reviewed By: int3

Differential Revision: D17357741

fbshipit-source-id: 4513c6d3c545619ef3fbe122bad00f836b0e7556