Make tracking of locations written by invokes more precise
Summary:
Up to now, if an invocation was known to write to any field that was also read in an optimized methods, then that invocation was treated as a general memory barrier, and no information was carried over.

For invocations where more specific barrier information is known, e.g. which exact (non-volatile) fields an invocation may write to, then that information is now used to clear out known value information more selectively, increasing the overall effectiveness of CSE slightly.

Also, giving a few more available bits to tracking locations.

Reviewed By: int3

Differential Revision: D17412887

fbshipit-source-id: 4b1a4d5874d8fdcd7eb1736bccc3f474be3393b9