implement references by reflection
Summary:
# Concealing references to classes

There are some findings by Lauren in our SDK regarding rare and unique library class references. This makes the adversary easy to locate a class even after its name is obfuscated. In order to tackle this problem, we can reference to this class and its members only by reflection. And only store them as Objects.

This diff does...
1. all references to a target class to object (handles renaming correctly with `TypeRefUpdater`)
2. Whenever it sees a field or method access with a reference to a target class, use reflection to call it. This also automatically fixes the issue that the type system won't allow you call an arbitrary instance method on Object.

With RESL, the strings will also be encrypted which provides better security.

Reviewed By: int3, Feng23

Differential Revision: D17555271

fbshipit-source-id: 067dec6d7ff8c7b05e9b4497d5fca1ecbdfdf73e