Replace apply_modifiers with a simple check for the rule type
Summary:
The current behavior is bizarre -- it doesn't apply keep modifiers to methods / fields in superclasses, but does mark them as `has_keep`. Other parts of the code assume that if `m_keep` is set, then the keep modifiers are set too, so this is just broken. I'm not sure why it was implemented this way to begin with.

I see no reason to have separate `apply_modifiers` boolean when we can simply look at the rule type and always apply the keep modifiers if we are processing a keep rule, so that's what this diff does.

Reviewed By: ssj933

Differential Revision: D17806231

fbshipit-source-id: b4d44ac28a59db5b536813686b6f763d451c9014