Compute fixpoint
Summary:
Consider the following method.

```
Ljava/util/ArrayList;.add:(Ljava/lang/Object;)Z
```

When this method gets invoked on an instance of `ArrayList`, and not some arbitrary subtype, then this invocation doesn't represent a memory barrier (`add` doesn't mutate any fields that non-framework code could access). CSE already takes this into account (for a set of such potentially safe virtual methods) when the exact receiver type of an invocation is known and matches.

The more often we know the exact receiver type, the better. In particular, we know it when the receiver value came from a `NEW_INSTANCE` instruction. However, the way store-to-load forwarding is implemented in CSE, it obscures the original source of a value, and thus CSE often cannot figure out the exact type of a forwarded value, and thus conservatively treats various method invocations as unsafe.

This diff runs CSE in a fixpoint loop. During every iteration, store-load forwarding gets a chance to create a more explicit link between a value usage and definition, thereby increasing the chance that fewer method invocations need to be treateds as memory barriers in subsequent iterations.

(I first tried to address this issue by tracking the original values through store-load forwarding, but this requires much more involved domains, dealing with sets of original values, growing fast in code complexity and adding runtime overhead even in the case when CSE won't be able to optimize anything in a method. Just running CSE to a fixpoint is much more straightforward, and results in possibly better overall performance.)

Reviewed By: amnn

Differential Revision: D17851552

fbshipit-source-id: a1238ed251e4ee110d9f5dca95fc76619d173104