proper fix for IPCP problem
Summary:
When building callgraph for IPCP, we are not including every methods, rather we go from root methods and methods that are true virtual, and expand the call graph. One bug is <clinit> usually don't have callsite in dex code, but they are also not true virtual, and not necessary root. So we may ignore <clinit> methods in our callgraph. So we fix this by including <clinit> in callgraph's root as well.

Will patch to unstable too when accepted

Reviewed By: int3

Differential Revision: D17960715

fbshipit-source-id: 8d1ce344e4157fac9945067a566e20bf7aad47d2