Analyze whether a java.lang.Object argument's type info is not used in the method
Summary:
It analyzes how `java.lang.Object` arguments are used in a method.
If a `java.lang.Object` argument is not possible to have any virtual call and casting in the method, and it's not escaping or only escaping as return value, that means its type information is never used in the method, it should be safe if we pass an Integer object instead of an Enum object after the enum-to-int transformation.

Previously we only analyze static methods, now all non-true-virtuals.

This diff should have no size impact and next diff will adopt the info in enum-to-integer transformation.

Reviewed By: int3

Differential Revision: D17754823

fbshipit-source-id: 2c3811c53703a8cc5243244d95fcb04fe7bee8f2