Introduce Litho class init counter
Summary:
This diff introduces a ClassInitCounter analysis, that tracks where an object of a specified class hierarchy is instantiated and then tracks all usages in methods and fields, and all escapes.

This includes all fields of the object either set or fetched, all methods invoked on the object, all methods that the object is passed to, and all external fields that the method is set as.

Escapes are control flow paths where the object value leaves the control of the originating method, including returns, set as the value of a field, or passed as a parameter to a method, or set into an array.
These are stored into an escapes object; however, the caller of the analysis may provide a set of methods that should be considered 'safe' escapes and any of these where the tracked object is a parameter are stored as safe escapes.

For example, with Litho components the method getOrCreateCommonProps statically takes an object as a parameter; however, in the context of Litho component inlining, this escape is perfectly expected and is not semantically losing control of the object's scope and life time.

Reviewed By: NTillmann

Differential Revision: D15658929

fbshipit-source-id: 3de08cfa38df368d6d041d77a8d4c662f7c6fa07