Introducing RemoveUninstantiablesPass
Summary:
A dedicated pass for simplifying away mentions of uninstantiable classes.  This
diff introduces the pass boilerplate, logic for iterating across methods,
simplification of the `instance-of` instruction, and an accompanying test.

This diff does *not* include the integration of this pass into any refigs. I am
planning on implementing the parts of the pass that are meant to replace
existing parts of other passes first.  After that I'll introduce a diff to add
this pass into the refigs whilst simultaneously removing the previous
implementation.

Reviewed By: int3

Differential Revision: D18377671

fbshipit-source-id: e4b2e2dd8c8a06a001ee3a882155153c33a16180