Fix accidental extra methodinline run
Summary:
Before this diff, PerfMethodInlinePass would run a regular IntraDexInline if no profile information was supplied. This is not intended. The issue is that we want PerfMethodInlinePass to be a purely additive change to the inliner right now, but with this bug, it was both additive and subtractive to our inlining decisions.

Fix by bailing out early if we have no method profiling information.

Reviewed By: ssj933

Differential Revision: D18429923

fbshipit-source-id: 97d41a9e9fec0dfc021008371a4c5e85739792a6