Fix CFGMutation on macOS
Summary:
macOS builds were failing because `CFGMutation` had a reference field and an explicitly defaulted move assignment operator.  The two are not compatible, and there was a warning enabled in the macOS toolchain that is not presented in the linux toolchain.  It, combined with `-Werror` caused the build to fail for macOS.

The solution was to delete the move assignment operator.

Reviewed By: NTillmann

Differential Revision: D18434293

fbshipit-source-id: a6b3b2c66505cb00e46cd45f07d4950bea5e5d1d