Working around a GCC bug in RemoveUninstantiablesPass
Summary:
I hit a GCC bug ([this one](https://gcc.gnu.org/bugzilla/show_bug.cgi?id=71546)) with one of my recent diffs.

Prior to GCC 8, the parser for lambda init captures fails to clear namespace context between initialisers.  This causes GCC to incorrectly complain that it can't find one of the captures.

The workaround is to change the order of initialisers so the unqualified one comes first.

[This godbolt](https://gcc.godbolt.org/z/Ddg4yq) contains a minimal example of the failure case, and [this godbolt](https://gcc.godbolt.org/z/3Eck6n) shows fix is to the workaround.

Reviewed By: int3

Differential Revision: D18436820

fbshipit-source-id: 56f4f02df9dd1f3699f979824927e068bbee0fb0