Extend CFGInliner with a plugin mechanism
Summary:
Inlining other than traditional methdo inlining can also benefit from using the CFGInliner, but may require additional fixups.
This plugin receives the caller and a copy of the callee cfg before and after registers are remapped in the callee, and can
make modifications according to the need of the additional inlining.

Reviewed By: justinjhendrick

Differential Revision: D18325115

fbshipit-source-id: 4fea22a02cd1aabdf54b8506cc32f0cc7ac17ab8